pyproject.toml
[project]
name = "devinotes-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.37",
]
